requirements.txt
Django==1.6.5
dj-database-url==0.3.0
dj-static==0.0.5
django-tastypie==0.11.1
django-toolbelt==0.0.1
google-api-python-client==1.2
gunicorn==18.0
httplib2==0.9
psycopg2==2.5.3
pystache==0.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.7.0
static==1.0.2
twilio==3.6.6
wsgiref==0.1.2
